Fix P2 Combo

+++ Zetris/GameHelper.cs
@@ -688,7 +688,7 @@ namespace Zetris {
                             Game.ReadInt32(new IntPtr(
                                 Game.ReadInt32(new IntPtr(
                                     Game.ReadInt32(new IntPtr(
-                                        Game.ReadInt32(new IntPtr(
+                                        Game.ReadByte(new IntPtr(
                                             0x140461B20
                                         )) + 0x378
                                     )) + 0x30
@@ -701,7 +701,7 @@ namespace Zetris {
                         ret = Game.ReadInt32(new IntPtr(
                             Game.ReadInt32(new IntPtr(
                                 Game.ReadInt32(new IntPtr(
-                                    Game.ReadInt32(new IntPtr(
+                                    Game.ReadByte(new IntPtr(
                                         0x140598A28
                                     )) + 0x140
                                 )) + 0x48
@@ -714,7 +714,7 @@ namespace Zetris {
                     case 0:
                         ret = Game.ReadInt32(new IntPtr(
                             Game.ReadInt32(new IntPtr(
-                                Game.ReadInt32(new IntPtr(
+                                Game.ReadByte(new IntPtr(
                                     0x140598A20
                                 )) + 0x38
                             )) + 0x3DC
@@ -724,7 +724,7 @@ namespace Zetris {
                     case 1:
                         ret = Game.ReadInt32(new IntPtr(
                             Game.ReadInt32(new IntPtr(
-                                Game.ReadInt32(new IntPtr(
+                                Game.ReadByte(new IntPtr(
                                     0x140598A28
                                 )) + 0x38
                             )) + 0x3DC
